custom-elements/custom-element-lifecycle/types-of-callbacks/entered-view-callback.html

<!DOCTYPE html>
<html>
<head>
<title>enteredView callback of a custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="enteredView callback must be enqueued whenever custom element is inserted into a document and this document has a browsing context">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    var callbackCalled = false;
    
    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback  = function () {
        callbackCalled = true;
    }
    
    assert_false(callbackCalled, 'Callback enteredView was called after prototype creation '
            + 'but should not');
    
    var XFoo = doc.register('x-foo-1', {prototype: proto});
    
    assert_false(callbackCalled, 'Callback enteredView was called but should not');
    
    var xfoo = new XFoo();

    assert_false(callbackCalled, 'Callback enteredView was called');
    
}, 'Test enteredView callback when custom element is instantiating in JavaScript. '
    + 'Document has no browsing context');


test(function() {
    var doc = newHTMLDocument();
    
    var callbackCalled = false;
    
    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalled = true;
    }
    
    assert_false(callbackCalled, 'Callback enteredView was called after prototype creation '
            + 'but should not');
    
    var XFoo = doc.register('x-foo-2', {prototype: proto});
    
    assert_false(callbackCalled, 'Callback enteredView was called but should not');
    
    doc.body.innerHTML = '<x-foo-2></x-foo-2>';

    assert_false(callbackCalled, 'Callback enteredView was called');
    
}, 'Test enteredView callback when custom element is instantiating in HTML'
    + 'Document has no browsing context');


test(function() {
    var doc = newHTMLDocument();
    
    var callbackCalled = false;
    
    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalled = true;
    }
    
    assert_false(callbackCalled, 'Callback enteredView was called after prototype '
            + 'creation but should not');
        
    doc.body.innerHTML = '<x-foo-3></x-foo-3>';
    
    assert_false(callbackCalled, 'Callback enteredView was called but should not');
    
    var XFoo = doc.register('x-foo-3', {prototype: proto});

    assert_false(callbackCalled, 'Callback enteredView was called');
    
}, 'Test enteredView callback when custom element is instantiating in HTML before '
    + 'registration of a custom element. Document has no browsing context');
    

testInIFrame('../../resources/x-foo.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    
    var callbackCalled = false;
    
    var proto = Object.create(HTMLElement.prototype);
    proto.enteredViewCallback = function () {
        callbackCalled = true;
    }
    
    assert_false(callbackCalled, 'Callback enteredView was called after prototype '
            + 'creation but should not');
            
    var XFoo = doc.register('x-foo', {prototype: proto});

    assert_true(callbackCalled, 'Callback created was not called');

}, 'Test enteredView callback. Document has browsing context');

    
</script>
</body>
</html>
